+++ .vscode/setting.json
@@ -1,3 +1,8 @@
 {
-	"editor.tabSize": 2
+	"editor.tabSize": 2,
+	"editor.defaultFormatter": "esbenp.prettier-vscode",
+  "editor.formatOnSave": true,
+  "editor.codeActionsOnSave": {
+    "source.fixAll.eslint": true
+  }
 }
