+++ frontend/src/pages/login.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import {View, Text} from "react-native";
+
+const Login = () => {
+	return (
+		<View>
+			<Text>ログイン画面</Text>
+		</View>
+	)
+}
+
+export default Login;
